agents/test-agent/requirements.txt
coinbase-agentkit>=0.7.0
python-dotenv>=1.0.0
